Move list into new TodoList component
