+++ src/App.jsx
@@ -1,31 +1,11 @@
 import * as React from "react";
-
-const todoList = [
-  {
-    id: 1,
-    title: "Complete assignment",
-  },
-  {
-    id: 2,
-    title: "Attend mentor session",
-  },
-  {
-    id: 3,
-    title: "Read related articles",
-  },
-];
+import TodoList from "./TodoList"; //Import TodoList component
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((item) => (
-          <li key={item.id}>
-            <span>{item.title}</span>
-          </li>
-        ))}
-      </ul>
+      <TodoList />
     </div>
   );
 }
